Starts service provider refactor
