--- src/Definition/ClosureDefinition.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace League\Container\Definition;
-
-/**
- * @deprecated will be removed in favour of `CallableDefinition`
- */
-class ClosureDefinition extends CallableDefinition
-{
-
-}
