+++ src/Reflector/ClassReflector.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace League\Container\Reflector;
+
+class ClassReflector
+{
+
+}
