--- src/ServiceProvider.php
@@ -1,38 +1,0 @@
-<?php
-
-namespace League\Container;
-
-abstract class ServiceProvider implements ContainerAwareInterface
-{
-    use ContainerAwareTrait;
-
-    /**
-     * @var array
-     */
-    protected $provides = [];
-
-    /**
-     * Use the register method to register items with the container via the
-     * protected $this->container property or the `getContainer` method
-     * from the ContainerAwareTrait.
-     *
-     * @return void
-     */
-    abstract public function register();
-
-    /**
-     * Returns a boolean if checking whether this provider provides a specific
-     * service or returns an array of provided services if no argument passed.
-     *
-     * @param  string $alias
-     * @return boolean|array
-     */
-    public function provides($alias = null)
-    {
-        if (! is_null($alias)) {
-            return (in_array($alias, $this->provides));
-        }
-
-        return $this->provides;
-    }
-}
