+++ src/ServiceProvider/ServiceProviderInterface.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace League\Container\ServiceProvider;
+
+interface ServiceProviderInterface extends ContainerAwareInterface
+{
+    /**
+     * Returns a boolean if checking whether this provider provides a specific
+     * service or returns an array of provided services if no argument passed.
+     *
+     * @param  string $alias
+     * @return boolean|array
+     */
+    public function provides($alias = null);
+
+    /**
+     * Use the register method to register items with the container via the
+     * protected $this->container property or the `getContainer` method
+     * from the ContainerAwareTrait.
+     *
+     * @return void
+     */
+    public function register();
+}
